+++ common/src/main/java/net/mehvahdjukaar/polytone/tabs/CreativeTabsModifiersManager.java
@@ -16,6 +16,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.CreativeModeTabs;
+import net.minecraft.world.item.DyeColor;
 
 import java.util.*;