+++ common/src/main/java/net/mehvahdjukaar/polytone/texture/DayTimeTexture.java
@@ -5,4 +5,8 @@ public interface DayTimeTexture {
     boolean polytone$usesDayTime();
 
     void polytone$setUsesDayTime(boolean usesWorldTime);
+
+    int polytone$getDayDuration();
+
+    void polytone$setDayDuration(int duration);
 }