+++ neoforge/src/main/java/net/mehvahdjukaar/polytone/neoforge/PlatStuffImpl.java
@@ -1,4 +1,4 @@
-package net.mehvahdjukaar.polytone.forge;
+package net.mehvahdjukaar.polytone.neoforge;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.mehvahdjukaar.polytone.mixins.forge.*;
@@ -10,11 +10,13 @@ import net.minecraft.client.color.item.ItemColor;
 import net.minecraft.client.color.item.ItemColors;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.multiplayer.SessionSearchTrees;
+import net.minecraft.client.particle.ParticleProvider;
 import net.minecraft.client.renderer.DimensionSpecialEffects;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
+import net.minecraft.core.particles.ParticleType;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
@@ -56,6 +58,10 @@ public class PlatStuffImpl {
         PolytoneForge.bus.addListener(eventConsumer);
     }
 
+    public static ParticleProvider<?> getParticleProvider(ParticleType<?> type) {
+        return ((ParticleEngineAccessor) Minecraft.getInstance().particleEngine)
+                .getProviders().get(BuiltInRegistries.PARTICLE_TYPE.getKey(type));
+    }
 
     public static BlockColor getBlockColor(BlockColors colors, Block block) {
         return ((BlockColorsAccessor) colors).getBlockColors().get(block);